GenPoem/backend/requirements.txt
Flask
flask-socketio
openai
textblob
emotion_analysis
google-generativeai

